requirements.txt
aiofiles==0.6.0
aiohttp==3.6.3
Async-OAuthlib==0.0.9
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
discord-ext-ipc==2.0.0
discord.py==1.7.2
DiscordUtils==1.2.9.2
dpytools==0.17.4b0
Flask==1.1.2
Flask-Caching==1.10.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
laftel==1.0.0
lavalink==3.1.4
MarkupSafe==1.1.1
multidict==4.7.6
oauthlib==3.1.0
Pillow==8.2.0
priority==1.3.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PUBGpy==1.0
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
python-dotenv==0.17.1
Quart==0.14.1
Quart-Discord==2.1.4
randomvalue==111
requests==2.25.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
wavelink==0.9.9
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.5.1
youtube-dl==2021.5.16
